# perf(ext/streams): optimize streams