… cli/tests/unit/streams_test.ts