@marcosc90 2023-09-23
> This is awesome.
> 
> Can you see if this test passes w/1000000 writes? It was too slow to pass for me before your PR:
> 
> https://github.com/denoland/deno/blob/main/cli/tests/unit/streams_test.ts#L252

Just rebased `main`, I'll try it out.