@marcosc90 2023-09-23
@mmastrac the optimizations are much better than I thought.

This PR with: `1_000_000`
```
readableStreamVeryLargePackets ... ok (235ms)
```

In `main` it's been **3 minutes** and still going, I'll update once it finishes...