@mmastrac 2023-09-23
I tested this patch with my web-stream serving benchmark and it's ~2% faster vs main. Great work!

57171 -> 58495 for large chunk streams

63183 -> 63183 for small chunk streams